[SAFRAN-926] Updated the cardinality of the identifier attribute of MediaType and updated the Validator service
